refactor: 프로젝트에 맞게 클래스 패키지 변경
