+++ backend/src/main/java/com/dwbh/backend/dto/LoginRequest.java
@@ -1,7 +1,6 @@
-package com.dwbh.backend.security.dto;
+package com.dwbh.backend.dto;
 
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
 public class LoginRequest {
